+++ src/main/java/subway/auth/ui/AuthController.java
@@ -1,16 +1,14 @@
 package subway.auth.ui;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import subway.auth.application.AuthService;
 import subway.auth.dto.TokenRequest;
 import subway.auth.dto.TokenResponse;
-import subway.auth.exception.InvalidMemberException;
+import subway.member.exception.InvalidMemberException;
 import subway.auth.infrastructure.JwtTokenProvider;
 
 @RestController
