+++ src/main/java/subway/favorite/application/FavoriteService.java
@@ -2,7 +2,7 @@ package subway.favorite.application;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import subway.auth.exception.FavoriteNotFoundException;
+import subway.favorite.exception.FavoriteNotFoundException;
 import subway.favorite.dao.FavoriteDao;
 import subway.favorite.domain.Favorite;
 import subway.favorite.dto.FavoriteRequest;
