+++ src/main/java/subway/favorite/exception/FavoriteNotFoundException.java
@@ -1,4 +1,4 @@
-package subway.auth.exception;
+package subway.favorite.exception;
 
 public class FavoriteNotFoundException extends NullPointerException {
 }
