+++ src/main/java/subway/line/exception/SectionNotFoundException.java
@@ -1,4 +1,4 @@
-package subway.auth.exception;
+package subway.line.exception;
 
 public class SectionNotFoundException extends NullPointerException{
 }
