+++ src/main/java/subway/member/dao/MemberDao.java
@@ -7,7 +7,7 @@ import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import subway.auth.exception.InvalidMemberException;
+import subway.member.exception.InvalidMemberException;
 import subway.member.domain.Member;
 
 import javax.sql.DataSource;
