+++ src/main/java/subway/member/exception/InvalidMemberException.java
@@ -1,4 +1,4 @@
-package subway.auth.exception;
+package subway.member.exception;
 
 public class InvalidMemberException extends IllegalArgumentException {
 }
